Feature: Replace 'double' with 'decimal' type handling in BaseFormRequest and AutoCrud for improved data validation

+++ src/Http/BaseFormRequest.php
@@ -73,8 +73,8 @@ abstract class BaseFormRequest extends FormRequest
             case 'number':
                 $fieldRules[] = 'integer';
                 break;
-            case 'double':
-                $fieldRules[] = 'double';
+            case 'decimal':
+                $fieldRules[] = 'decimal';
                 break;
             case 'select':
                 if (isset($field['options'])) {
